+++ public/index.html
@@ -41,6 +41,7 @@
                             data-size="large"
                             data-logo_alignment="left"
                         ></div>
+                        <button type="button" onclick="googleCallback()">Click Me!</button>
                     </div>
                 </div>
             </div>
