+++ public/js/googleAuth.js
@@ -1,11 +1,15 @@
 //  GOOGLE FUNCTIONS
 //callback function that is used when sign-in button is clicked.
 async function googleCallback(googleResponse) {
-    let validateResponse = await validateGoogleToken(googleResponse); //validates that signin happened, and that id_token is accepted
-    createSessionCookie('google-session-token', googleResponse.credential, 7); //Creates a cookie that is active for 7 days
+    // let validateResponse = await validateGoogleToken(googleResponse); //validates that signin happened, and that id_token is accepted
+    // createSessionCookie('google-session-token', googleResponse.credential, 7); //Creates a cookie that is active for 7 days
     let urlResponse = await getAuthorizationURL(); //get authorization url from server.
-    let jsonResponse = await urlResponse.json(); //converts response to json
-    window.location.href = jsonResponse.url; //opens the URL that enables user to
+    if(urlResponse != false) { //if redirect url is not returned, the server failed and client cannot be redirected.
+        let jsonResponse = await urlResponse.json(); //converts response to json
+        window.location.href = jsonResponse.url; //opens the URL that enables user to
+    } else {
+        window.alert("Error in redirection to authorization procedure!"); //alert user that server failed and redirection did not happen
+    }
 }
 
 //function that validates idToken with Google
@@ -14,17 +18,22 @@ async function validateGoogleToken(googleResponse) {
         method: 'POST',
         mode: 'no-cors',
         headers: { 'Content-type': 'application/json; charset=UTF-8' },
-        body: googleResponse.credential //
+        body: googleResponse.credential
     });
     return response;
 }
 
 //function gets authorizationURL from server.
 async function getAuthorizationURL() {
-    let redirectURL = await fetch(`http://localhost:3000/authorizationRedirect`, {
-        method: 'GET'
-    });
-    return redirectURL;
+    try{
+        let redirectURL = await fetch(`http://localhost:3000/authorizationRedirect`, {
+            method: 'GET'
+        });
+        return redirectURL;
+    } catch {
+        return false;
+    }
+    
 }
 
 //  COOKIE-HANDLE
